+++ src/components/tomogram/main.tsx
@@ -21,7 +21,12 @@ import { PlotContainer } from "components/visualisation/plotContainer";
 import { Motion } from "components/motion/motion";
 import { useCallback } from "react";
 import { client, prependApiUrl } from "utils/api/client";
-import { TomogramData, BaseProcessingJobProps, DataConfig, AutoProcSchema } from "schema/interfaces";
+import {
+  TomogramData,
+  BaseProcessingJobProps,
+  DataConfig,
+  AutoProcSchema,
+} from "schema/interfaces";
 import { CTF } from "components/ctf/ctf";
 import { components } from "schema/main";
 import { ProcessingTitle } from "components/visualisation/processingTitle";